refactor: reduce search method size
